+++ packages/MJCoreEntitiesServer/src/index.ts
@@ -28,6 +28,7 @@ export * from './custom/MJAIAgentSessionBridgeEntityServer.server';
 export * from './custom/MJAIAgentSessionBridgeParticipantEntityServer.server';
 export * from './custom/MJAIRemoteBrowserProviderEntityServer.server';
 export * from './custom/MJAISkillPermissionEntityServer.server';
+export * from './custom/MJAISkillEntityServer.server';
 export * from './custom/json-schema-lite';
 export * from './custom/IntegrationLLMPKCallback';
 export * from './custom/util';